Refactor MPT sequential.
