+++ mp2-common/src/eth.rs
@@ -10,11 +10,12 @@ use ethers::{
     },
 };
 use rlp::{Encodable, Rlp, RlpStream};
+use serde::{Deserialize, Serialize};
 #[cfg(feature = "ci")]
 use std::env;
 use std::{array::from_fn as create_array, sync::Arc};
 
-use crate::{mpt_sequential::bytes_to_nibbles, rlp::MAX_KEY_NIBBLE_LEN, utils::keccak256};
+use crate::{mpt_sequential::utils::bytes_to_nibbles, rlp::MAX_KEY_NIBBLE_LEN, utils::keccak256};
 /// A wrapper around a transaction and its receipt. The receipt is used to filter
 /// bad transactions, so we only compute over valid transactions.
 pub struct TxAndReceipt(Transaction, TransactionReceipt);
@@ -243,7 +244,7 @@ pub struct ProofQuery {
     pub(crate) slot: StorageSlot,
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub enum StorageSlot {
     /// simple storage slot like a uin256 etc that fits in 32bytes
     /// Argument is the slot location in the contract
